+++ Helpers/Extend.js
@@ -1,115 +1,67 @@
-import { isFunction, isArray } from 'front-library/Helpers/Type';
+import { isFunction, isArray, isPlainObject } from 'front-library/Helpers/Type';
 
+
+/**
+ * Extend function from jQuery 3
+ */
 function _extend() {
-    var options,
-        name,
-        src,
-        copy,
-        copyIsArray,
-        clone,
-        target = arguments[0] || {},
+    var options, name, src, copy, copyIsArray, clone,
+        target = arguments[ 0 ] || {},
         i = 1,
         length = arguments.length,
-        deep = true,
-        class2type = {},
-        core_toString = class2type.toString,
-        core_hasOwn = class2type.hasOwnProperty;
-
-    'Boolean Number String Function Array Date RegExp Object Error'
-        .split(' ')
-        .forEach(function(name) {
-            class2type['[object ' + name + ']'] = name.toLowerCase();
-        });
-
-    function type(obj) {
-        if (obj == null) {
-            return String(obj);
-        }
-        // Support: Safari <= 5.1 (functionish RegExp)
-        return typeof obj === 'object' || typeof obj === 'function'
-            ? class2type[core_toString.call(obj)] || 'object'
-            : typeof obj;
-    }
-
-    function isPlainObject(obj) {
-        // Not plain objects:
-        // - Any object or value whose internal [[Class]] property is not "[object Object]"
-        // - DOM nodes
-        // - window
-        if (type(obj) !== 'object' || obj.nodeType || obj === window) {
-            return false;
-        }
-
-        // Support: Firefox <20
-        // The try/catch suppresses exceptions thrown when attempting to access
-        // the "constructor" property of certain host objects, ie. |window.location|
-        // https://bugzilla.mozilla.org/show_bug.cgi?id=814622
-        try {
-            if (
-                obj.constructor &&
-                !core_hasOwn.call(obj.constructor.prototype, 'isPrototypeOf')
-            ) {
-                return false;
-            }
-        } catch (e) {
-            return false;
-        }
-
-        // If the function hasn't returned already, we're confident that
-        // |obj| is a plain object, created by {} or constructed with new Object
-        return true;
-    }
+        deep = false;
 
     // Handle a deep copy situation
-    if (typeof target === 'boolean') {
+    if ( typeof target === 'boolean' ) {
         deep = target;
-        target = arguments[1] || {};
-        // skip the boolean and the target
-        i = 2;
+
+        // Skip the boolean and the target
+        target = arguments[ i ] || {};
+        i++;
     }
 
     // Handle case when target is a string or something (possible in deep copy)
-    if (typeof target !== 'object' && !isFunction(target)) {
+    if ( typeof target !== 'object' && !isFunction( target ) ) {
         target = {};
     }
 
-    // extend jQuery itself if only one argument is passed
-    if (length === i) {
-        target = this;
-        --i;
-    }
 
-    for (; i < length; i++) {
+    for ( ; i < length; i++ ) {
+
         // Only deal with non-null/undefined values
-        if ((options = arguments[i]) != null) {
-            // Extend the base object
-            for (name in options) {
-                src = target[name];
-                copy = options[name];
+        if ( ( options = arguments[ i ] ) != null ) {
 
+            // Extend the base object
+            for ( name in options ) {
+                copy = options[ name ];
+
+                // Prevent Object.prototype pollution
                 // Prevent never-ending loop
-                if (target === copy) {
+                if ( name === "__proto__" || target === copy ) {
                     continue;
                 }
+
                 // Recurse if we're merging plain objects or arrays
-                if (
-                    deep &&
-                    copy &&
-                    (isPlainObject(copy) || (copyIsArray = isArray(copy)))
-                ) {
-                    if (copyIsArray) {
-                        copyIsArray = false;
-                        clone = src && isArray(src) ? src : [];
+                if ( deep && copy && ( isPlainObject( copy ) ||
+                    ( copyIsArray = isArray( copy ) ) ) ) {
+                    src = target[ name ];
+
+                    // Ensure proper type for the source value
+                    if ( copyIsArray && !isArray( src ) ) {
+                        clone = [];
+                    } else if ( !copyIsArray && !isPlainObject( src ) ) {
+                        clone = {};
                     } else {
-                        clone = src && isPlainObject(src) ? src : {};
+                        clone = src;
                     }
+                    copyIsArray = false;
 
                     // Never move original objects, clone them
-                    target[name] = _extend(deep, clone, copy);
+                    target[ name ] = _extend( deep, clone, copy );
 
-                    // Don't bring in undefined values
-                } else if (copy !== undefined) {
-                    target[name] = copy;
+                // Don't bring in undefined values
+                } else if ( copy !== undefined ) {
+                    target[ name ] = copy;
                 }
             }
         }